Refactoring

+++ Entities/Concrete/UserAddress.cs
@@ -10,6 +10,6 @@ namespace Entities.Concrete
         public int Id { get; set; }
         public int UserId { get; set; }
         public string AddressText { get; set; }
-        public int CityId { get; set; }
+        public byte CityId { get; set; }
     }
 }
